Struct tracing_actix_web::RequestId

pub struct RequestId(_);

A unique identifier generated for each incoming request.

Extracting a RequestId when the TracingLogger middleware is not registered will result in an internal server error.

Usage

use actix_web::get;
use tracing_actix_web::RequestId;
use uuid::Uuid;

#[get("/")]
async fn index(request_id: RequestId) -> String {
  format!("{}", request_id)
}

#[get("/2")]
async fn index2(request_id: RequestId) -> String {
 let uuid: Uuid = request_id.into();
  format!("{}", uuid)
}

Optionally, using the uuid_v7 feature flag will allow RequestId to use UUID v7 instead of the currently used UUID v4.

However, the uuid crate requires a compile time flag uuid_unstable to be passed in RUSTFLAGS="--cfg uuid_unstable" in order to compile. You can read more about it here.